[#81] Support for overwriting clangd options
- introduce ClangdOptionsDefaults interface and remove DS trick
- add missing copyrights
