+++ bundles/org.eclipse.cdt.lsp.clangd/OSGI-INF/org.eclipse.cdt.lsp.clangd.BuiltinClangdOptionsDefaults.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.cdt.lsp.clangd.BuiltinClangdOptionsDefaults">
+   <property name="service.ranking" type="Integer" value="0"/>
+   <service>
+      <provide interface="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults"/>
+   </service>
+   <implementation class="org.eclipse.cdt.lsp.clangd.BuiltinClangdOptionsDefaults"/>
+</scr:component>
