+++ examples/org.eclipse.cdt.lsp.examples.preferences/OSGI-INF/org.eclipse.cdt.lsp.examples.preferences.MyClangdOptionsDefaults.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.cdt.lsp.examples.preferences.MyClangdOptionsDefaults">
    <property name="service.ranking" type="Integer" value="100"/>
-   <property name="isDefaultOptions" type="Boolean" value="true"/>
    <service>
-      <provide interface="org.eclipse.cdt.lsp.clangd.ClangdOptions"/>
+      <provide interface="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults"/>
    </service>
    <implementation class="org.eclipse.cdt.lsp.examples.preferences.MyClangdOptionsDefaults"/>
 </scr:component>
